requirements.txt
netmiko
GitPython
python-dotenv
pythonping
